commits

+++ src/main/java/com/electra/web/model/Payment.java
@@ -13,6 +13,6 @@ public class Payment{
     private Long id;
     private Double amount;
     private Date paymentDate;
-    private int customerId;
-    private int orderId;
+    private Long customerId;
+    private Long orderId;
 }
